- removed static_cast with proper code where possible ( firebird/common.h, soci-db2.h, soci-firebird.h). - restored soci-config.h with pragma message that header is obsolete if anyone still uses it - removed MSVC wrappers and restored original code

+++ include/soci/firebird/soci-firebird.h
@@ -219,7 +219,7 @@ protected:
     long long rowsAffectedBulk_; // number of rows affected by the last bulk operation
 
     virtual void exchangeData(bool gotData, int row);
-    virtual void prepareSQLDA(XSQLDA ** sqldap, int size = 10);
+    virtual void prepareSQLDA(XSQLDA ** sqldap, short size = 10);
     virtual void rewriteQuery(std::string const & query,
         std::vector<char> & buffer);
     virtual void rewriteParameters(std::string const & src,
