+++ src/backends/mysql/common.cpp
@@ -68,11 +68,11 @@ void soci::details::mysql::parse_std_tm(char const *buf, std::tm &t)
     details::mktime_from_ymdhms(t, year, month, day, hour, minute, second);
 }
 
-char * soci::details::mysql::quote(MYSQL * conn, const char *s, int len)
+char * soci::details::mysql::quote(MYSQL * conn, const char *s, size_t len)
 {
     char *retv = new char[2 * len + 3];
     retv[0] = '\'';
-    int len_esc = mysql_real_escape_string(conn, retv + 1, s, len);
+    int len_esc = mysql_real_escape_string(conn, retv + 1, s, static_cast<unsigned long>(len));
     retv[len_esc + 1] = '\'';
     retv[len_esc + 2] = '\0';
