+++ src/backends/mysql/vector-use-type.cpp
@@ -78,7 +78,7 @@ void mysql_vector_use_type_backend::pre_use(indicator const *ind)
                     std::vector<std::string> &v = *pv;
 
                     buf = quote(statement_.session_.conn_,
-                        v[i].c_str(), static_cast<int>(v[i].size()));
+                        v[i].c_str(), v[i].size());
                 }
                 break;
             case x_short:
